requirements.txt
asgiref==3.3.1
boto3==1.17.7
botocore==1.20.7
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
cloudinary==1.24.0
codegen==1.0
confusable-homoglyphs==3.2.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.5
Cython==0.29.21
databases==0.4.1
decorator==4.4.2
defusedxml==0.7.0rc2
dj-database-url==0.5.0
Django==3.1.6
django-allauth==0.44.0
django-cloudinary-storage==0.3.0
django-cors-headers==3.7.0
django-email-utils==1.0.0
django-environ==0.4.5
django-etc==1.3.0
django-filter==2.4.0
django-heroku==0.3.1
django-hitcount==1.3.3
django-registration==3.1.1
django-rest-auth==0.9.5
django-rest-authemail==2.0.6
django-rest-email-auth==3.0.3
django-rest-knox==4.1.0
django-rest-passwordreset==1.1.0
django-rest-swagger==2.2.0
django-swagger-ui==0.1.11
django-templated-mail==1.1.1
django-utils-six==2.0
Django-Verify-Email==0.0.5
django-versatileimagefield==2.0
django-viewtracker==1.2
djangorestframework==3.12.2
djangorestframework-filters==0.11.1
djangorestframework-simplejwt==4.6.0
djoser==2.1.0
drf-yasg==1.20.0
dynaconf==3.1.2
form==0.0.1
gunicorn==20.0.4
hdbscan==0.8.27
hvac==0.10.8
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.3
jmespath==0.10.0
joblib==1.0.1
knox==0.1.11
loguru==0.5.3
Markdown==3.3.3
MarkupSafe==1.1.1
numpy==1.20.1
oauthlib==3.1.0
openapi-codec==1.3.2
packaging==20.9
patterns==0.3
Pillow==8.1.0
psycopg2==2.8.6
psycopg2-binary==2.8.6
pycparser==2.20
PyJWT==2.0.1
pyparsing==2.4.7
python-dateutil==2.8.1
python-magic==0.4.18
python3-openid==3.2.0
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
ruamel.yaml==0.16.12
s3transfer==0.3.4
scikit-learn==0.24.1
scipy==1.6.0
shortuuid==1.0.1
simplejson==3.17.2
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==4.0.3
SQLAlchemy==1.3.23
sqlparse==0.4.1
threadpoolctl==2.1.0
uritemplate==3.0.1
urllib3==1.26.3
validators==0.18.2
views==0.3
whitenoise==5.2.0
